requirements.txt
streamlit==1.14.0
rembg==2.0.28
Pillow==9.2.0
requests==2.28.1
pandas==1.5.3
numpy==1.23.4
streamlit-lottie==0.1.17
